Remove Eigen, start reorganising the c++ file
The Eigen library has been removed since that it can install via aptitude.
Change the gitignore in order to exclude the bak file from emacs.
Add a pro file in order to use qmake and qt library.
Add a file in order to perform some color conversion.
Add a file in order to perform some segmentation.
Add the main file in order to make the detection.